refactor(MongoClient): make connection use operations/connect.js
