+++ lib/mongo_client.js
@@ -3,6 +3,7 @@
 const ChangeStream = require('./change_stream');
 const Db = require('./db');
 const EventEmitter = require('events').EventEmitter;
+const executeOperation = require('./operations/execute_operation');
 const executeLegacyOperation = require('./utils').executeLegacyOperation;
 const inherits = require('util').inherits;
 const MongoError = require('./core').MongoError;
@@ -10,8 +11,7 @@ const deprecate = require('util').deprecate;
 const WriteConcern = require('./write_concern');
 
 // Operations
-const connectOp = require('./operations/mongo_client_ops').connectOp;
-const validOptions = require('./operations/mongo_client_ops').validOptions;
+const ConnectOperation = require('./operations/connect');
 const closeOperation = require('./operations/mongo_client_ops').closeOperation;
 
 /**
@@ -192,16 +192,13 @@ Object.defineProperty(MongoClient.prototype, 'writeConcern', {
  * @return {Promise<MongoClient>} returns Promise if no callback passed
  */
 MongoClient.prototype.connect = function(callback) {
-  // Validate options object
-  const err = validOptions(this.s.options);
-
   if (typeof callback === 'string') {
     throw new TypeError('`connect` only accepts a callback');
   }
 
-  return executeLegacyOperation(this, connectOp, [this, err, callback], {
-    skipSessions: true
-  });
+  const operation = new ConnectOperation(this);
+
+  return executeOperation(this, operation, callback);
 };
 
 MongoClient.prototype.logout = deprecate(function(options, callback) {
