+++ lib/operations/connect.js
@@ -1,6 +1,8 @@
 'use strict';
 
+const os = require('os');
 const OperationBase = require('./operation').OperationBase;
+const defineAspects = require('./operation').defineAspects;
 const Aspect = require('./operation').Aspect;
 const deprecate = require('util').deprecate;
 const Logger = require('../core').Logger;
@@ -133,8 +135,39 @@ const validOptionNames = [
   'retryWrites',
   'useNewUrlParser',
   'useUnifiedTopology',
-  'serverSelectionTimeoutMS'
+  'serverSelectionTimeoutMS',
+  'useRecoveryToken',
+  'autoEncryption'
 ];
+
+const ignoreOptionNames = ['native_parser'];
+const legacyOptionNames = ['server', 'replset', 'replSet', 'mongos', 'db'];
+
+// Validate options object
+function validOptions(options) {
+  const _validOptions = validOptionNames.concat(legacyOptionNames);
+
+  for (const name in options) {
+    if (ignoreOptionNames.indexOf(name) !== -1) {
+      continue;
+    }
+
+    if (_validOptions.indexOf(name) === -1) {
+      if (options.validateOptions) {
+        return new MongoError(`option ${name} is not supported`);
+      } else {
+        console.warn(`the options [${name}] is not supported`);
+      }
+    }
+
+    if (legacyOptionNames.indexOf(name) !== -1) {
+      console.warn(
+        `the server/replset/mongos/db options are deprecated, ` +
+          `all their options are supported at the top level of the options object [${validOptionNames}]`
+      );
+    }
+  }
+}
 
 const LEGACY_OPTIONS_MAP = validOptionNames.reduce((obj, name) => {
   obj[name.toLowerCase()] = name;
@@ -142,20 +175,15 @@ const LEGACY_OPTIONS_MAP = validOptionNames.reduce((obj, name) => {
 }, {});
 
 class ConnectOperation extends OperationBase {
-  constructor(mongoClient, err) {
+  constructor(mongoClient) {
     super();
 
     this.mongoClient = mongoClient;
-    this.err = err;
-  }
-
-  static get aspects() {
-    return new Set([Aspect.SKIP_SESSIONS]);
   }
 
   execute(callback) {
     const mongoClient = this.mongoClient;
-    const err = this.err;
+    const err = validOptions(mongoClient.s.options);
 
     // Did we have a validation error
     if (err) return callback(err);
@@ -166,6 +194,7 @@ class ConnectOperation extends OperationBase {
     });
   }
 }
+defineAspects(ConnectOperation, [Aspect.SKIP_SESSION]);
 
 function addListeners(mongoClient, topology) {
   topology.on('authenticated', createListener(mongoClient, 'authenticated'));
@@ -245,6 +274,7 @@ function connect(mongoClient, url, options, callback) {
     if (_finalOptions.socketTimeoutMS == null) _finalOptions.socketTimeoutMS = 360000;
     if (_finalOptions.connectTimeoutMS == null) _finalOptions.connectTimeoutMS = 30000;
     if (_finalOptions.retryWrites == null) _finalOptions.retryWrites = true;
+    if (_finalOptions.useRecoveryToken == null) _finalOptions.useRecoveryToken = true;
 
     if (_finalOptions.db_options && _finalOptions.db_options.auth) {
       delete _finalOptions.db_options.auth;
@@ -433,7 +463,46 @@ function createTopology(mongoClient, topologyType, options, callback) {
     }
 
     assignTopology(mongoClient, newTopology);
-    callback(null, newTopology);
+    if (options.autoEncryption == null) {
+      callback(null, newTopology);
+      return;
+    }
+
+    // setup for client side encryption
+    let AutoEncrypter;
+    try {
+      AutoEncrypter = require('mongodb-client-encryption').AutoEncrypter;
+    } catch (err) {
+      callback(
+        new MongoError(
+          'Auto-encryption requested, but the module is not installed. Please add `mongodb-client-encryption` as a dependency of your project'
+        )
+      );
+
+      return;
+    }
+
+    const MongoClient = loadClient();
+    let connectionString;
+    if (options.autoEncryption.extraOptions && options.autoEncryption.extraOptions.mongocryptURI) {
+      connectionString = options.autoEncryption.extraOptions.mongocryptURI;
+    } else if (os.platform() === 'win32') {
+      connectionString = 'mongodb://localhost:27020/?serverSelectionTimeoutMS=1000';
+    } else {
+      connectionString = 'mongodb://%2Ftmp%2Fmongocryptd.sock/?serverSelectionTimeoutMS=1000';
+    }
+
+    const mongocryptdClient = new MongoClient(connectionString, { useUnifiedTopology: true });
+    mongocryptdClient.connect(err => {
+      if (err) return callback(err, null);
+
+      const mongoCryptOptions = Object.assign({}, options.autoEncryption, {
+        mongocryptdClient
+      });
+
+      topology.s.options.autoEncrypter = new AutoEncrypter(mongoClient, mongoCryptOptions);
+      callback(null, newTopology);
+    });
   });
 }
